+++ ts/packages/knowPro/src/index.ts
@@ -14,8 +14,6 @@ export * from "./propertyIndex.js";
 export * from "./timestampIndex.js";
 export * from "./serialization.js";
 export * from "./dateTimeSchema.js";
-export * from "./searchSchema.js";
-export * from "./searchTranslator.js";
 export * from "./searchQueryTranslator.js";
 export * from "./textLocationIndex.js";
 export * from "./messageIndex.js";
